[GLib] Add Travis CI configuration for building Arrow GLib with Meson

+++ ci/travis_before_script_c_glib.sh
@@ -27,11 +27,16 @@ if [ $TRAVIS_OS_NAME == "osx" ]; then
   export PKG_CONFIG_PATH=$PKG_CONFIG_PATH:/usr/local/opt/libffi/lib/pkgconfig
 fi
 
+if [ $BUILD_SYSTEM = "meson" ]; then
+  source $TRAVIS_BUILD_DIR/ci/travis_install_conda.sh
+  pip install meson ninja
+fi
+
 gem install test-unit gobject-introspection
 
 if [ $TRAVIS_OS_NAME == "osx" ]; then
   sudo env PKG_CONFIG_PATH=$PKG_CONFIG_PATH luarocks install lgi
-else
+elif [ $BUILD_SYSTEM = "autotools" ]; then
   git clone \
     --quiet \
     --depth 1 \
@@ -45,41 +50,57 @@ else
   luarocks install lgi
 fi
 
-go get github.com/linuxdeepin/go-gir-generator || :
-pushd $GOPATH/src/github.com/linuxdeepin/go-gir-generator
-rm lib.in/gio-2.0/gdk_workaround.go
-mv lib.in/gio-2.0/config.json{,.orig}
-sed \
-  -e 's/\("Settings",\)/\/\/ \1/g' \
-  -e 's/\("SettingsBackend",\)/\/\/ \1/g' \
-  lib.in/gio-2.0/config.json.orig > lib.in/gio-2.0/config.json
-mv Makefile{,.orig}
-sed -e 's/ gudev-1.0//' Makefile.orig > Makefile
-mkdir -p out/src/gir/gudev-1.0
-make build copyfile
-mkdir -p $GOPATH/bin/
-cp -a out/gir-generator $GOPATH/bin/
-cp -a out/src/gir/ $GOPATH/src/gir/
-popd
+if [ $BUILD_SYSTEM = "autotools" ]; then
+  go get github.com/linuxdeepin/go-gir-generator || :
+  pushd $GOPATH/src/github.com/linuxdeepin/go-gir-generator
+  rm lib.in/gio-2.0/gdk_workaround.go
+  mv lib.in/gio-2.0/config.json{,.orig}
+  sed \
+    -e 's/\("Settings",\)/\/\/ \1/g' \
+    -e 's/\("SettingsBackend",\)/\/\/ \1/g' \
+    lib.in/gio-2.0/config.json.orig > lib.in/gio-2.0/config.json
+  mv Makefile{,.orig}
+  sed -e 's/ gudev-1.0//' Makefile.orig > Makefile
+  mkdir -p out/src/gir/gudev-1.0
+  make build copyfile
+  mkdir -p $GOPATH/bin/
+  cp -a out/gir-generator $GOPATH/bin/
+  cp -a out/src/gir/ $GOPATH/src/gir/
+  popd
+fi
 
 pushd $ARROW_C_GLIB_DIR
-
-./autogen.sh
 
 export PKG_CONFIG_PATH=$PKG_CONFIG_PATH:$ARROW_CPP_INSTALL/lib/pkgconfig
 export LD_LIBRARY_PATH=$LD_LIBRARY_PATH:$ARROW_CPP_INSTALL/lib
 
-CONFIGURE_OPTIONS="--prefix=$ARROW_C_GLIB_INSTALL"
-if [ $TRAVIS_OS_NAME != "osx" ]; then
-  CONFIGURE_OPTIONS="$CONFIGURE_OPTIONS --enable-gtk-doc"
+if [ $BUILD_SYSTEM = "autotools" ]; then
+  ./autogen.sh
+
+  CONFIGURE_OPTIONS="--prefix=$ARROW_C_GLIB_INSTALL"
+  if [ $TRAVIS_OS_NAME != "osx" ]; then
+    CONFIGURE_OPTIONS="$CONFIGURE_OPTIONS --enable-gtk-doc"
+  fi
+
+  CONFIGURE_OPTIONS="$CONFIGURE_OPTIONS CFLAGS=-DARROW_NO_DEPRECATED_API"
+  CONFIGURE_OPTIONS="$CONFIGURE_OPTIONS CXXFLAGS=-DARROW_NO_DEPRECATED_API"
+
+  ./configure $CONFIGURE_OPTIONS
+
+  make -j4
+  make install
+else
+  MESON_OPTIONS="--prefix=$ARROW_C_GLIB_INSTALL"
+  MESON_OPTIONS="$MESON_OPTIONS -Denable_gtk_doc=true"
+  mkdir -p build
+  env \
+    CFLAGS="-DARROW_NO_DEPRECATED_API" \
+    CXXFLAGS="-DARROW_NO_DEPRECATED_API" \
+    meson build $MESON_OPTIONS
+  pushd build
+  ninja
+  ninja install
+  popd
 fi
-
-CONFIGURE_OPTIONS="$CONFIGURE_OPTIONS CFLAGS=-DARROW_NO_DEPRECATED_API"
-CONFIGURE_OPTIONS="$CONFIGURE_OPTIONS CXXFLAGS=-DARROW_NO_DEPRECATED_API"
-
-./configure $CONFIGURE_OPTIONS
-
-make -j4
-make install
 
 popd
